requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
pyserial==3.5
pydantic==2.7.1
pytest==8.2.0
python-socketio==5.11.2
